refactor: make fadeSlide reusable Signed-off-by: Philippe Martin <phmartin@redhat.com>
